requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
boto3==1.35.92
botocore==1.35.92
Brotli==1.1.0
celery==5.4.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2023.10
crispy-tailwind==1.0.3
cryptography==39.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.0.3
django-allauth==0.61.1
django-browser-reload==1.16.0
django-celery-results==2.5.1
django-crispy-forms==2.1
django-dbbackup==4.2.1
django-debug-toolbar==4.2.0
django-htmx==1.18.0
django-storages==1.14.4
fonttools==4.53.1
gunicorn==21.2.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
kombu==5.3.7
oauthlib==3.2.2
packaging==23.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
pycparser==2.21
pydyf==0.11.0
PyJWT==2.6.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
redis==5.0.8
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.10.4
six==1.16.0
sqlparse==0.4.3
tinycss2==1.3.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.14
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
